third_party/WebKit/LayoutTests/media/controls/overflow-menu-hide-on-scroll.html

<!DOCTYPE html>
<title>Ensure overflow menu hides on scroll.</title>
<script src='../../resources/testharness.js'></script>
<script src='../../resources/testharnessreport.js'></script>
<script src='../media-controls.js'></script>
<video controls></video>
<div style='height: 1000px'>foo</div>
<div style='height: 1000px' tabindex=0>bar</div>
<script>
async_test(t => {
  const video = document.querySelector('video');
  video.src = '../content/test.ogv';
  const track = video.addTextTrack('captions');

  video.onloadeddata = t.step_func(() => {
    const menu = overflowMenu(video);

    singleTapOnControl(overflowButton(video), t.step_func(() => {
      assert_not_equals(getComputedStyle(menu).display, 'none');

      // Scrolling hides the menu.
      document.querySelectorAll('div')[1].focus();
      setTimeout(t.step_func_done(() => {
        assert_equals(getComputedStyle(menu).display, 'none');
      }));
    }));
  });
});
</script>
